chore(sink): split engine-turn tests into a submodule (repo-shape length cap)
